automation-service/requirements.txt
# Automation Service Dependencies

# LLM Providers (same as ai-service)
anthropic==0.42.0
google-generativeai==0.8.3
groq==0.11.0

# Database
asyncpg==0.29.0

# Redis
redis==5.0.1

# Email — sending
sendgrid==6.11.0

# Email — receiving / parsing
# (imaplib is stdlib, no extra package needed)
beautifulsoup4==4.12.3       # parse HTML email bodies

# Attachment parsing
pdfplumber==0.11.4           # PDF text extraction
python-docx==1.1.2           # Word document text extraction
openpyxl==3.1.5              # Excel text extraction
Pillow==11.0.0               # Image handling

# Scheduling
APScheduler==3.10.4

# HTTP
httpx==0.27.2

# Data
pydantic==2.9.2
python-dateutil==2.8.2
json-repair==0.30.3

# Utilities
pytz==2024.2
